--- frontend_client/laptop_store_client_page/src/hooks/table/index.ts
@@ -1,12 +1,0 @@
-import { Options } from './types';
-
-const useTable = <TData>({ data, columns }: Options<TData>) => {
-    const getHeaderGroup = () => columns.map((item) => item.header);
-    const getRows = () => {};
-    const getContext = () => {};
-    const getAllSelected = () => {};
-    const getIsSelected = () => {};
-    const toggleAllSelected = (selected: boolean) => {};
-    const toggleSelected = (selected: boolean) => {};
-    return {};
-};
