+++ frontend_client/laptop_store_client_page/src/libs/utilities/index.ts
@@ -1,0 +1,1 @@
+export type { JwtType } from './types';
