+++ frontend_client/laptop_store_client_page/src/libs/utilities/types.ts
@@ -1,6 +1,6 @@
-export type JwtType = {
+export interface JwtType {
     sub: string;
     iat: number;
     iss: string;
     exp: number;
-};
+}
