--- frontend_client/laptop_store_client_page/src/libs/utils/types.ts
@@ -1,6 +1,0 @@
-export type JwtType = {
-    sub: string;
-    iat: number;
-    iss: string;
-    exp: number;
-};
